SIMPLE-5 moves impl from include to source
